THE WORLD'S BEST KEPT C SECRETS, PART 12: ----------------------------------------
The functions isalpha, isdigit, ..., isspace, tolower, and toupper
are ONLY defined on EOF (typically -1) and unsigned characters.
Read the manual pages from various OSs -- it's in there.

It is wrong to call these with signed characters as argument; the
type "char" can be, and often is, signed.  Please don't write code
like

     void frob (char *s)
     {
       while (isspace (*s))       /* WRONG! */
         s++;
       /* ... */
     }

Instead write something like

     void frob (char *s)
     {
       while (isspace ((unsigned char)*s))   /* Right */
         s++;
       /* ... */
     }

Note: the above is true even if you don't like it.  Some people have
gotten terrible upset and defensive when they have had this problem
pointed out to them.  That does not help.

FYI, all this makes a serious difference on Solaris, HP-UX and maybe
other OSs.

+++ src/dialogs/dialog-plugin-manager.c
@@ -116,7 +116,7 @@ refresh_right_frame (PluginData *pd, PluginManager *pm)
 	if (!pd)
 		str = g_strdup (_("File size:"));
 	else
-		str = g_strdup_printf (_("File size: %d bytes"), sbuf.st_size);
+		str = g_strdup_printf (_("File size: %ld bytes"), (long)sbuf.st_size);
 	gtk_label_set_text (GTK_LABEL (pm->size_lbl), str);
 	g_free (str);
